admin两个项目改造完毕

+++ admin-server/service/order/order-core/src/main/java/com/kakrolot/service/order/dao/po/QueryOrder.java
@@ -32,7 +32,7 @@ public class QueryOrder extends BasePO {
 
     private Long orderNum;
 
-    private Long orderAmount;
+    private BigDecimal orderAmount;
 
     private Long userId;
